macos: build and link against libedit
libreadline is GPL v3. Python supports linking against libedit, which has
a more permissive license. So, this commit teaches the build system to
build libedit and to link the readline extension against it.

Doing this with our custom-built ncurses resulted in a failure to find
the terminfo database. The workaround is to build/link against the
system ncurses. I'm not sure if this is ideal, as it leaves a dependency
on libraries in /usr/lib. But it is the easiest way to make the "readline"
module work. I'll have to research how all this stuff works and see if
we can ship ncurses safely on macOS.