macos: build LLVM/Clang for macOS
We need llvm-ar to build CPython with --with-lto. Apple's system LLVM/Clang
toolchain doesn't include llvm-ar. So we'll need to build it from source.

Building LLVM/Clang will also ensure that Python is built with a modern
toolchain. Otherwise we're at the whim of whatever toolchain is installed
on the system.

This commit only builds LLVM/Clang - it doesn't yet hook it up to building
of other packages.