macos: build LLVM/Clang 6
Because of LLVM intrinsics mismatch with Rust 1.31. As the added docs say,
we will want the ability to vary the used toolchain so we can produce
distributions that match downstream requirements. Because we are a
sister project of the PyOxidizer project, we will change things to match
the Rust toolchain until a more proper solution is in place.