windows: build a static _ssl extension
We still dynamically link against libcrypto and libssl. To fix that,
it looks like we'll need to build OpenSSL from source. I'm leaving that
for another day.

This required a couple of new hacks to build files to get working.
There is some property files magic going on in OpenSSL land...