windows: build _ctypes extension
This requires a minimal preprocessor check change to detect building
a built-in ctypes extension while still building a shared library.

This likely isn't a common configuration. But it is currently our
configuration (until we make libpython static).

With this change, we now have no .pyd files in the produced
distribution! I think we're still producing e.g. a tcltk .pyd.
But I don't care about that at this juncture.