windows: collect build artifacts in distribution
This /might/ be enough to get the build artifacts recombined into
a new binary.

This change also introduces the PYTHON.json file. It seems more
robust to have a file describing the distribution than to have
downstream consumers resort to heuristics to parse the distribution's
contents. If done properly, this should enable us to distribute
non-CPython distributions with minimal effort.

We'll need to teach the Linux and macOS distributions to produce
this file. We'll do that later.