linux: add unaccounted Modules/*.o files to core objects
Some object files under Modules/ are part of libpython and
not extension modules. We need to add these to the core object
files list so they are always compiled.

This fixes downstream linker errors due to missing symbols.
But we have run-time errors due to missing extension entries.
This is because we're not parsing Setup.dist for these
extensions.