linux: parse Setup.dist for extensions
We now correctly identify nearly every Modules/*.o file as being
associated with an extension. However, PyOxidizer still has
problems at run-time because we're not picking up the
special extensions that are part of the core build system
and have no entry in Setup.dist/Setup.local. This includes
the _warnings extension. More work is needed.