linux: parse config.c.in for in-core extensions
config.c.in defines extensions that aren't built using the
Setup.* files. These extensions are part of the core object
files.

This commit teaches things to parse config.c.in for those
entries so we can annotate the extensions in PYTHON.json
accordingly.

With this in place, PyOxidizer seems to be working on Linux
using PYTHON.json as the sole source of distribution/build
data!