linux: use libedit instead of libreadline
readline is GPL 3.0 licensed. libedit has a more permissive
license.

This commit teaches the Linux build to use libedit instead of
libreadline.

Due to the change, we get a warning at run-time about the
inability to locate the termcap database. There is still more
work to do here. But at least the GPL 3.0 issue is behind us.