linux: compile readline extension against both libreadline and libedit
Now that we have support for variants in the PYTHON.json file,
we can start to put it to use.

This commit teaches the Linux build system to build the
readline extension with both libreadline and libedit.

In order to accomplish this, we hacked up static-modules to
support custom syntax which is translated to custom build
rules. We then taught the distribution "parsing" code to
handle extension variants.

This is all a bit hacky. But it gets the job done.

As part of this, Python's configure and the installable
distribution now revert back to readline instead of libedit.
While there are GPLv3 issues, I think this is acceptable. We
have to pick 1 and readline is the only supported "backend"
on Linux. So I think readline makes sense over libedit.