linux: support building Python and dependencies with musl
I believe this works!

We had to use --disable-shared in various places because we can't
build shared libraries with musl. In theory this should be possible.
I suspect the musl-clang wrapper is passing conflicting
compiler/linker flags or something.

C++ compilation doesn't work with musl. So we disabled C++
compilation in some components. This would be a reasonable change
otherwise as we should never use the C++ bits.

Some hacks were required here and there. For example, OpenSSL
insists on looking for kernel headers, which we don't have. So
we had to disable some OpenSSL functionality to get things to
build. Disabling secure memory is sounds a bit scary. I left a
TODO to track investigating that.