linux: add license into to PYTHON.json
Now that we've annotated licenses for libraries, it is time to add
that annotation to the PYTHON.json file.

We bump the version of PYTHON.json to 2 and define fields to hold
license info.

We annotate libraries with their library names so we can find
licenses from link requirements.

Then we hook it all together and add the license annotations to
PYTHON.json and add the license files to the distribution.