windows: annotate license metadata
Like we did for Linux and macOS, we add licenses metadata to the
PYTHON.json for Windows builds.

This required a bit more effort because some libraries are built into
the extension itself and there's no easy way to automatically associate
libraries with extensions since it is all baked into Visual Studio
Project files.