linux: support LibreSSL
The impetus for this is that I couldn't get OpenSSL 1.1.1c building
against musl libc. But LibreSSL is a compelling replacement for OpenSSL
and I think it is a good idea to support it. We'll want this eventually
for BSD support anyway.

This commit teaches the Linux build system to build LibreSSL and to
have Python use it. AFAICT things "just work."