linux: build libX11 from source and use for tk build
This does a lot of things. Most notably the subject line.
We also split up tcl and tk into separate packages.

Chasing the long tail of libX11's dependencies was a PITA.

We still don't build the _tkinter extension. I think we still
need to work on making libX11 a static library.