linux: build _tkinter extension
The last commit did the heavy lifting of getting tcl/tk
built from source.

This commit puts the final touches on things so we can enable the
_tkinter extension module. We had to tweak the tcl/tk build
settings a bit. We also defined missing license annotations
for those packages.

With this commit `import tkinter` works from a standalone
executable! Although there may be some paths leaking into
the binary creating problems for compatibility on different
distros. But at least we're past the milestone of building
the extension!