utils: increase compression level for distribution
This reduces the total archive size a bit:

before: 42,518,472
after:  24,246,166

The only change that should affect consumers is the window size. Before,
the window size was 2^22. It is now 2^27. 128 MB is a bit large. But the
target user is developers, who should have plenty of memory. So I think
this is still acceptable. We can always decrease the window size if
people complain.