unix: port build_cpython() to build_environment()
This required implementing some new methods on the build environment.
As part of this, we also removed some now unused functions.

And with this, we can now build CPython using the build_environment()
API, opening the door to using the same building code without Docker
and supporting other operating systems!