unix: support compiling clang for macOS
Recent commits have started to make the UNIX build aware of different
platforms. This commit starts the concrete journey to teach the UNIX
build to work on macOS.

The build environment on macOS is quite different, as it doesn't use
Docker. And Clang in particular is sufficiently different as to
justify a separate build script from Linux. We could probably consolidate
these scripts. But I like keeping the shell code simple if possible.