windows: support building shared library distributions
Up until this point we produced a special Python distribution that
was statically linked. This can be useful in some scenarios. But
it can be very limiting because Python extensions need to be
built specifically against this distribution. This rules out
reusing existing binary extensions published on PyPI and
externalizes a lot of hardship on downstream consumers.

This commit teaches our build system to produce a more traditional
Python distribution. We have a pythonXY.dll for the core library
and .pyd files for extensions. There are also .dlls for various
dependencies instead of static libraries.

Closes #22.