windows: support building with PGO
We had this partially implemented but it wasn't quite working. It
probably got bit rotted once we discovered that static builds and PGO
didn't work too well due to linking issues downstream. With shared
builds, we're consuming DLLs and there shouldn't be linking issues.
So it should be safe to use PGO optimized binaries.

This commit also contains a minor fixup or two for Windows builds.