windows: build tkinter in shared builds
Building the _tkinter extension module is much, much simpler in
shared builds. This commit teaches the build to build this extension
as well as install its resource files when performing a more
traditional shared library build.

As part of this, we needed to teach license mapping to map multiple
downloads to incorporate licenses.