windows: use proper openssl archives
This regressed a while ago and I never noticed, likely because
I had the old tar files sitting around in the build/ directory.

As part of this, we also check the future result to catch errors.
Otherwise things continue to build even if the archive fails to
extract and we get a weird error about missing files!