unix: obtain distribution metadata by running built interpreter
This is a more robust method of obtaining distribution metadata.

We implement support for populating bytecode fields. We also
change how the platform tag and abi tag is obtained to use this
method as well.

macOS is still broken since we need to implement a build environment
method for it. I'll do this later, once Linux is fully supporting
the new fields.