cpython: enable _testcapi extension
This is a high risk change.

On Linux, I'm no longer getting errors while building this extension
with Python 3.7. If it works, it works. Let's go with it.

I suspect it is still broken somewhere, especially since I'm pretty
sure I ran into errors with this extension last night. If it is,
we should guard the disabling of the extension more tightly so
it is available in configurations where it works.

This reduces the number of test failures from 51 to <10 on Linux
Python 3.7.