unix: change tcl install layout
Tcl seems to look for tcl resources under lib/* by default. This
layout also agrees with my Linux distribution. And tcl's configure
doesn't seem to allow rooting tcl resources outside of --libdir.

So, this commit installs resources to lib/ instead of lib/tcl/ and
updates the distribution metadata accordingly.

Because we no longer have an isolated tcl directory, we needed to
invent a new field to hold the set of sub-directories holding
tcl resources. This is a bit hacky. But it gets the job done.

This drops the number of CPython test failures from ~9 to ~5.

Closes #34.