windows: install test extensions
This is needed to prevent a bunch of test harness failures.

And we even had a TODO tracking this.

The extensions don't have their metadata annotated, just like
on UNIX. Also, the official Python installers seem to install the
extensions as well. So we are slowly converging on parity with
them.

Unfortunately, we can only do this on shared distributions, as building
in static mode yields a ton of missing symbol errors. This is probably
fixable. But will require more work. We can do that as a follow-up.