windows: support building CPython 3.8
The remaining work to build 3.8 was mostly related to building
libffi. This is definitely one of the wonkier dependencies
to build from source, especially since the source release
published by Python lacks a commit necessary to make the
build not error. I wasted a few hours tracking this down. *sigh*.

After a few false starts, including reimplementing aspects of the
`prepare_libffi.bat` script, I gave up and just used it verbatim.
We may come to regret this shortcut to support static builds,
which don't yet work for non-ffi reasons. But at least
`--profile shared` works!

libffi requiring cygwin is a bit annoying. But what are you going
to do?

Once libffi was being built, we encountered a packaging error
due to a new test extension module. So that change snuck in as
well.