windows: build libffi separately
Building libffi takes a while and iterating on Python's build
while having to wait on libffi is annoying.

This commit refactors the libffi building so it occurs as its
own step in the build, similarly to how OpenSSL is built.

As part of this, we need to hack up python.props to reference the
new location of the libffi assets.