unix: build a dynamic libpython unless targeting musl
python-build-standalone had its origins in supporting a single
file, statically linked Python interpreter. So its defaults were
to not produce a libpython shared library.

Fast forward several months. The PYTHON.json file format is now
sufficiently defined to the point where it is possible for a
single distribution to contain both a static and shared library
and for downstream consumers to be smart about which one they
consume. If a consumer wants a static libpython, they can statically
link it. If a consumer wants a shared libpython, they can use the
shared library version. If a consumer wants a custom libpython,
they can grab the set of object files and library dependencies that
are needed and produce a custom one, linking statically or
dynamically as they desire.

Furthermore, the lack of a dynamic libpython was non-standard and
caused issues in some environments. It would be nice for our Python
distributions to be more usable.

This commit teaches the UNIX build system to produce a dynamic
libpython unless building for musl.

As part of making macOS work, I discovered that libz.1.dylib was being
searched for in a non-sensical directory. So that was fixed. I also
discovered that CPython's build system links module libraries
against the `python` executable instead of the library. Why, I'm
not sure.

To make the produced distributions more portable, we need to hack
up RPATH in binaries linking against libpython so they can find
the library relative to the executable: no LD_LIBRARY_PATH mucking
required.

As part of this work, I noticed we weren't populating shared_lib and
static_lib in the build info metadata. So I added this as well.

Closes #44.