unix: enable tcl/tk to use CoreFoundation
Before, they couldn't find CoreFoundation because the custom Clang we use
doesn't know about the default search paths for macOS frameworks. So we define
the include and framework path so it can find and use CoreFoundation.

As part of this, the number of compiler warnings exploded. So we had to
suppress more warnings to keep things in check.