unix: build and use libncursesw so we have Unicode support
This required changing ncurses to build with wide character support,
which appends a `w` to the library name and include path.

libedit was a bit stubborn, as it didn't look for `libncursesw` by
default. And the inline patch is horrible because we couldn't easily
run autoconf in the old Debian Jessie build environment.

The Python build also needed some tweaks to pick up the new library
and include path. We also had to add some preprocessor defines so we
enable wide support in curses.

I verified that `curses.unget_wch` now exists.

Closes #47. Originally reported at
https://github.com/indygreg/PyOxidizer/issues/250.