windows: build statically linked libffi
Up until this point, static builds for Python 3.8 have been broken
because of issues linking libffi. This commit unbusts things.

We teach libffi to build a static library. This entailed changing
some defines so declspec(dllexport) isn't used. We also had to opt in
to a statically linked CRT. And disable a shared library.

With the static libffi.lib being produced, we had to teach our hacks
to the static build configuration to link it.

With all that in place, we're able to build a statically linked Python
3.8!