macos: build gettext from source
Python's configure searches for a usable libintl, which is provided by gettext.

If you have a libintl shared library on your system, configure will pick it
up and add a dynamic library dependency on it, making binaries non-portable.

To work around this, we build gettext from source and have Python statically
link against its libintl.

The configure workaround feels a bit shady to me. I'm not sure why the test
program finding the textdomain symbol fails to build. But we can link against
libintl.a and Python functions calling into it seem to work. So who knows.