unix: define DEFAULT_SYSROOT when building on macOS
Without this, our clang on Big Sur is unable to find libSystem.dylib and other
system libraries. My guess is there is some default macOS SDK search path logic
that is failing to account for macOS 11. Probing for the default SDK
directory as used by Xcode seems to result in things working.