ci: use sccache
Building in CI takes forever, mainly due to compiling Clang. Let's
use sccache to speed things up.

As part of this, we also adopt some refinements to locating the macOS
SDK. This is because the 11.0 SDK doesn't play well with CPython 3.8 or
Clang.

After this commit, I'm able to build CPython 3.8 and 3.9 in CI in ~40
minutes with a populated sccache.