ci: add Linux build
This required various refactoring to support sccache.

We now look for an sccache executable in the repository root directory. If
present, we copy it into the build environment.

We also tweak handling of environment variables so variables used by
sccache are proxied into the container environment.

The use of sccache is necessary to keep build times reasonable. Without
sccache, we hit job execution timeouts. With it, builds take like 60
minutes. Still slow. But better than not finishing.

There are definitely some optimizations still on the table, such as
building the toolchains in a single job and caching toolchain builds.
That would drastically speed up job times.