rust: add binary to validate tarballs
I've been wanting to add stronger validation of the produced archives. We
started to do this in Python a few commits ago. However, we want to do some
low-level validation and doing this in Python will be brittle. So this
commit introduces a Rust binary. The first thing this binary does is validate
zstd compressed tarballs and examine mach-o files for allowed library
dependencies. It's a start.

As part of this, we update CI to build the binary and run it to perform
validation of the built tarball distribution.