unix: initial support for Apple arm64 builds
This commit teaches our build system to perform cross-compiling on
Apple, with initial support for targeting Apple ARM64.

There's a lot going on here. I attempted to comment things liberally
inline because cross-compiling is hard. The high-level gist of it
is we now pass `-arch arm64` into compiler and linker flags when
targeting Apple arm64.

CPython's build system doesn't officially support cross builds on
macOS. So we had to hack up various configure checks to get things to
work. This was likely the most time intensive portion of the
implementation.

I don't have access to an Apple M1 device, so I am unable to verify
this works as expected. But we do seem to produce arm64 binaries, so
I think we're on the right track.

I'm sure there's plenty of minor fixes that need to be implemented.
And we don't yet have CI for the new target.

It's also worth mentioning that with basic cross build support implemented,
we can likely get iPhone builds working with minimal effort. I'm not
sure though.