unix: always add Apple SDK arguments
Before, we defined an environment variable for the Apple SDK path and
added -isysroot and other flags in each build script. When building for
Apple, we should be able to universally add -isysroot and other flags
to ensure the appropriate SDK is targeted. This will enable us to more
easily target other SDKs, such as iOS, without having to modify multiple
build scripts.

This also has the benefit that our compiler invocations are explicit
about which SDK they use. Before, we'd pick up the SDK using Clang's
heuristics. This was probably OK since we built Clang from source. But
in cases where we wanted to target an SDK, it wouldn't do the correct
thing before.