unix: initial support for aarch-64-apple-ios target
The build fails on ncurses. But at least we get the plumbing in place to
target the iOS platform.

As part of this, I learned that open source Clang doesn't support the
arm64e target. Seriously. I guess we'll need to teach our build system
to run the Apple Clang present on the host when targeting iOS. We can
do this in a later commit.