ci: split toolchain build into own job
Previously, each job built the toolchain from source. This took a long
time and was redundant.

This commit splits building of the toolchain into its own job and has
each Python build job download and use the pre-built toolchain.

The ultimate goal is to enable caching of the toolchain archive so we
can skip building it entirely. But we'll do that in a separate commit.
The goal of this commit is to isolate the building of the toolchain to
its own job.