unix: reference libffi properly when building ctypes extension on Linux
The -I path here was out of date and didn't exist. And we failed to pull
in the additional link library location for the static library. Let's
fix both of those things.

I don't believe this will have a meaningful impact on the build because
these paths/arguments are getting added via external/global flags. But
capturing them here is proper.