unix: define test extensions in static modules files
Previously, setup.py would build these extensions. When cross-compiling
on Apple, setup.py doesn't target the exact architecture we want it
to and our validation script would fail.

Having modules listed in static-modules is preferred for our project
because it gives us explicit control over how things are built and
doesn't rely on setup.py logic, which does some undesirable things
in situations. So we should have a policy of nerfing setup.py whenever
possible.

The one exception to using static-modules appears to be the _testcapi
extension. This extension needs to build against the public interface
and can't be built as a built-in. See the comment in the added
disabled-static-modules file.

I _think_ this will yield a working cross-build with validation in CI!