unix: make Docker image names variables
In order to support cross compiling, we'll be introducing new Docker
images containing cross toolchains. This means we need to make the
Docker image names variable.

This commit writes out the Docker image names to the target-specific
Makefile. We then use the variables when resolving dependencies.

In order to support using dynamic image names, we teach build.py to
accept the image name as an argument. We feed the image name from the
Makefile variable through to build.py.

There should be no net change in behavior as a result of this change.