unix: improve ncurses cross-compiling
Previously, cross-compiles only worked on Apple and x86_64 -> i686.
Upcoming commits to introduce ARM Linux cross-compiles uncovered more
issues.

This commit contains changes to work around those.

We have to save and restore CC to get the host build to use the
correct compiler.

We also need to munge the path to the host compiler to use the host
GCC present in the new cross-compile environment.

And we need to disable application stripping, since `strip` doesn't
recognize the target. This is probably the hackiest change. But I
couldn't figure out (only spent minimal effort though) on how to
get ncurses's build system to use the correct `strip` binary.