windows: link _ctypes to libffi in static builds
CPython 3.8 moved to an out-of-tree libffi. As far as I can tell, our
first official Python 3.8 static builds did not handle libffi correctly.

In 3.8 and 3.9 builds, we hacked up the Python project to link against
libffi.lib. This ensured the symbols were present at link time. However,
we didn't ship libffi.lib and the dependency of _ctypes on libffi.lib
wasn't captured in PYTHON.json. If we attempted to produce a new link
library from the object files + static libraries (such as what
PyOxidizer does), we'd get a linker error for missing libffi symbols.

This commit attempts to fix matters.

We now explicitly link the _ctypes extension to libffi.lib via
project files hackery. This dependency is captured in the metadata
so the link dependency is present in PYTHON.json. We also need to
manually copy libffi.lib into the build directory so it is
available.

Finally, we undo the dependency of pythoncore on libffi because I now
believe it is redundant with _ctypes. Although I'm not 100% confident
about this.

This should hopefully fix
https://github.com/indygreg/PyOxidizer/issues/360.