windows: annotate pathcch dependency for pythoncore
This dependency was added in Python 3.9 as part of
bpo-39401 / Git commit 6a65eba44bfd82ccc8bed4b5c6dd6637549955d5.

The backport to Python 3.8 introduced run-time sniffing of the
symbol to avoid the hard library dependency.

This commit adds pathcch as a dependency for pythoncore. No
attention is made to the Python version, effectively meaning we
now require Windows 8 since this library is only available on
Windows 8. I'm fine with not claiming support for Windows 7: that
makes our support story simpler.

This should fix issues with missing symbols for e.g.
PathCchCanonicalizeEx when relinking a new library from the PYTHON.json
metadata.