global: add Apple SDK metadata to PYTHON.json
https://github.com/indygreg/PyOxidizer/issues/373 has revealed issues
relinking the Python distributions on older Apple SDKs and Clang
toolchains. In order to help downstream consumers validate they are
using a sufficiently modern SDK and toolchain, this commit adds metadata
to the Python distribution to advertise information about the Apple
SDK used to build and its targeting settings.

As part of this, we teach the Rust validation code to ensure fields
are present on Apple triples.