unix: properly define PYTHON_ARCH for more triples
We were inserting malformed symlinks into the archive because we
didn't annotate all target triples. This commit hopefully fixes
that.

As part of debugging this, I also discovered that cross builds
were missing "d" in abiflags. This commit should also fix that.

To prevent this from happening again we teach the validation code to
ensure all symlink targets exist in the archive.