downloads: update tcl/tk to 8.6.12 / remove tix on macOS
As part of the upgrade, tix stopped compiling on macOS due to compiler
errors. What the root cause is I'm not sure. It is probably possible
to work around. However, as part of investigating this I noticed that
the official Python macOS pkg installers don't ship the tix tcl files.
And the `tkinter.tix` module has been deprecated since Python 3.6. So
I think it is acceptable to drop support for tix on macOS.