unix: always symlink libpythonX.Y.a
This was previously behind a conditional on having a binary suffix. In
reality the static library is always present and was previously being
duplicated as separate files instead of a symlink.

This commit should ensure we only have a single copy of the library
and that the `python/install/lib/libpythonX.Y.a` symlink always exists.

Closes #73.