ncurses: make build ABI-compatible with version 5
I built Mercurial on Mac with PyOxidizer. I then ran `hg commit -i` and quit without changing anything. That resulted in a segfault. It seems that the issue is an ABI compatibility between the ncurses version compiled against (6.1 or 6.2, I don't remember which) and the version dynamically linked on macOS. On Big Sur, the ncurses  version seems to be 5.4. Adding `--with-abi-version=5` to the `configure` script fixed the problem for me.