unix: install pip and setuptools from wheels
This commit refactors the install of pip and setuptools.

Previously, we installed both of these by invoking `setup.py` from the
host Python. This worked, but it was somewhat old school. And in the
case of cross-compiling, it installed the packages to the host Python,
not the target Python.

After this commit, we download wheels for both pip and setuptools and
the install invokes the host Python but runs the pip wheel to invoke
pip to install the packages. This is conceptually similar to how the
`ensurepip` module works.

The end result is that pip is used to install itself and setuptools
into the appropriate output directory. This works with both native
and cross builds.

Aside from minor changes to the directory layout, the end result is
the same as far as I can tell. But since cross-compiles now install
pip and setuptools correctly, this fixes #84.

Because pip now bootstraps self and this doesn't work on musl builds
without a patch, we had to move the patching of pip to before it
is invoked. We moved the patching of setuptools as well, because it
is related.

Windows builds were updated to use the wheel based pip and setuptools.
However, we don't support cross-compiling on Windows yet, so we
continue to use the built python.exe to install these wheels.