unix: change _dyld_shared_cache_contains_path logic
The commits to the last CPython 3.8 and 3.9 releases tracked by
https://bugs.python.org/issue44689 break our builds because they make
assumptions about the build environment that don't hold in our build
environment. We apply a patch to partially revert this upstream
change.

After this patch, our builds may not succeed or work correctly when
performed on a <11.0 SDK. The next patches will update the CI build
environment accordingly.