unix: attempt to work around permissions issue with wheels
CI is encountering a permissions issue with the wheel files. I can't
reproduce this locally. Might have something to do with umasks. I
dunno. Let's try to work around it by removing the source wheel
files, which shouldn't require write privileges on the file - only
the directory it is part of.