unix: remove support for building with LibreSSL
Python 3.10 goes all in on requiring OpenSSL 1.1.1 and it requires some
patches to build with LibreSSL due to newer functionality from 1.1.1 not
yet present in LibreSSL.

The main reason we introduced support for LibreSSL in the first place
was because we couldn't get an older version to compile against musl
libc. It appears the latest OpenSSL now compiles fine.

We were only using LibreSSL on the musl builds: everything else was
using OpenSSL. (Way back when we shipped LibreSSL as the default SSL
implementation but reverted this when it was clear CPython really
wanted OpenSSL.)

So the practical impact of this change is musl libc distributions
switch from LibreSSL to OpenSSL.