unix: stop building _gdbm extension on 3.10
With the switch from readline to libedit, _gdbm was our final GPL
licensed dependency on Linux. In my opinion, the benefits of removing
the last GPL/copyleft license from the distributions outweighs the
benefits that the _gdbm extension provides.

If someone complains, we could potentially provide GPL versions
of the distributions containing readline and _gdbm. But until
someone asks, let's avoid the complexity.