unix: upgrade llvm/clang to 13.0.0
Python 3.6+ is now a build dependency. On macOS, the system Python
should be sufficiently new. So the build should hopefully pick it
up. On Linux, our old Debian version only ships Python 3.4, so we
need to install a modern Python. We simply build it from source as
part of the clang build script.

A better solution would be to establish a dedicated "host Python"
artifact. This way we could use this Python build for both Clang
and the host Python used in cross-compiles. But the current solution
in this patch is much simpler.