unix: simplify configure nerfs during cross-compiling
All the override values here were the same for all triples. And forcing
CROSS_COMPILING in more scenarios in recent commits appears to have
tickled an issue on musl libc. So let's make the configs universal.

AFAICT the risk with the /dev/pt* checks is minimal, as Python's
`os.openpty()` should use a function from `pty.h` and not rely on
the device handles. But there is a risk we're doing something subtly
wrong.

TBH I'm not sure why device presence is a configure time check. I
would think this is the type of thing that should be a run-time
check since different environments have different devices. So I
suspect there's a CPython portability bug lingering in here...