unix: use sccache more during gcc build
I discovered this optimization as part of extracting the toolchain
bootstrap logic to a separate, not-yet-published project. Let's
"backport" it to python-build-standalone to hopefully realize a
build speedup.

In order to make this work, we had to move `sccache` to PATH. Well,
we strictly didn't need to. But I think that makes more sense.