rust: restructure GitHub artifact downloading
Before, we initiated the HTTP requests all at once. On slow network
connections or machines this resulted in timeouts due to some requests
not being serviced in a timely manner.

This commit refactors the code so requests are issued when their
future is waited on and not before. This should result in more
reliable fetching of artifacts.