rust: change Windows triple configuration
This fixes a discrepancy between what we were building in CI and
what we were releasing. The upload step should now not complain
about missing Windows artifacts.

A side-effect of this is we produce install only archives
for Windows static builds.

I'm not super thrilled about lying about the triple value. But
it works and keeps code simpler.