rust: change release artifact naming convention
We change the filename so the tag is placed after the Python versions.
e.g. `cpython-X.Y-<triple>-<build>-<tag>.tar.zst` ->
`cpython-X.Y+<tag>-<triple>-<build>.tar.zst`. This naming scheme is
better because the tag is effectively a version component and it
should be earlier in the filename to aid sorting.