rust: use object for elf parsing
This completes the transition from goblin to the object
crate for parsing binary files.

As part of this, we fix a bug handling big-endian ELF binaries,
as our new code for symbol reading should be endian aware.

Some of this code was borrowed from the linux-package-analyzer
crate. I wrote the code so there should be no copyright issue.