rust: support validating Mach-O symbol references against tbd files
macOS SDKs have TBD files, which are YAML files that define library
information, such as which symbols are exported.

We can use these TBD files to validate that symbol references in
our Mach-O files reference symbols that actually exist in the
targeted SDK version.

This commit implements that functionality.

There are a few issues with this:

1. Older SDKs don't have `SDKSettings.json` files. So our SDK search
   is failing.
2. The 10.9 SDK (which we still target) doesn't ship TBD files.

So in the current implementation we only validate down to the
10.14 SDK. This is not sufficient to detect the bug that #122
purportedly addresses.