rust: examine SDKs with an SDKSettings.plist
Building on top of the previous commit, this commit teaches our Apple
SDK validation to handle Apple SDKs with only an SDKSettings.plist, not
an SDKSettings.json. This enables us to parse macOS SDKs 10.10+.

10.9 SDKs still don't validate since they lack .tbd files.

With this change, I'm still not seeing any missing symbols. So
if #122 is fixing something, it must be with the 10.9 SDK or there
must be an error elsewhere, possibly in this validation code.