unix: stop using weak symbols with Python 3.8
The bane of weak symbols on macOS has come back to haunt us.
(See https://github.com/indygreg/PyOxidizer/issues/373 for previous
battles.)

In https://github.com/indygreg/python-build-standalone/pull/122
we tracked down a runtime failure to the fact that CPython 3.8
didn't properly backport weak symbol handling support. So, if
you build with a modern SDK targeting an older SDK (which we do as
of 63f13fb2c8c72038edf556b4413b6ec23894a7c6), the linker will
insert a weak symbol. However, CPython doesn't have the runtime
guards and will attempt to dereference it, causing a crash.

Up to this point, our strategy for handling this mess was to
stop using symbols on all Python versions when we found one to
be causing an issue. This was crude, but effective.

In recent commits, we implemented support for leveraging the
macOS SDK .tbd files for validating symbol presence. We can now
cross reference undefined symbols in our binaries against what
the SDKs tell us is present and screen for missing symbols. This
helps us detect strong symbols that aren't present on targeted
SDK versions.

For weak symbols, I'm not sure if we can statically analyze the
Mach-O to determine if a symbol is guarded. I _think_ the guard
is a compiler built-in and gets converted to a function call, or
maybe inline assembly. We _might_ have to disassemble if we wanted
to catch unguarded weakly referenced symbols. Yeah, no.

In this commit, we effectively change our strategy for weak symbol
handling.

Knowing that CPython 3.9+ should have guarded weak symbols everywhere,
we only ban symbol use on CPython 3.8, specifically x86-64 3.8 since
the aarch64 build targets macOS SDK 11, which has the symbols we
need.

We also remove the one-off validation check for 2 banned symbols. In
its place we add validation that only a specific allow list of weak
symbols is present on CPython 3.8 builds.

As part of developing this, I discovered that libffi was also using
mkostemp without runtime guards. I'm unsure if Python would ever call
into a function that would attempt to resolve this symbol. But if it
does it would crash on 10.9. So we disable that symbol for builds
targeting 10.9.