unix: use unified LLVM project source archive
Before this commit we'd been building LLVM+Clang by downloading separate
source archives and extracting them manually. LLVM has a unified source
archive with all the projects in it. A single file is easier to consume.
And the layout of the resulting source matches the Git repo layout,
which means we can easily substitute a Git checkout if we wanted.

A side-effect of this change is we have to use `LLVM_ENABLE_PROJECTS`
because of how LLVM's CMake build system works when the unified source
archive is used. This is probably for the best, as it is the preferred
way to build LLVM from source and it matches behavior of Git checkouts.

There should be no meaningful changes in the built Clang toolchain as a
result of this commit. I spot verified that the number of ninja jobs for
each stage matches what we did before and that the set of output files
in the toolchain archives are identical.

An interesting peculiarity is I noticed that we were not installing
libunwind despite extracting its archive. That might be an oversight,
but I think that's fine, as we probably don't want libunwind on Linux
anyway.